feat(entry-dialog): unify pushup + exercise entry into a single dialog
Replace the two separate dialogs (CreateEntryDialog for pushups,
EntryDialog for catalog exercises) and the per-category dashboard
sections with a single TrainingEntryDialog that picks the training type
via category + exercise dropdowns. Edit mode locks the pickers so an
entry cannot be moved between the legacy `pushups` and the new
`exerciseEntries` collections.

Wires the dashboard "Neu" button, stats-table create/edit, and
entries.store.createEntry to the new discriminated `kind` payload, and
removes the per-exercise cards from the dashboard. Carries over the
split minutes/seconds duration inputs from #306 into the unified dialog.
