+++ web/src/app/stats/components/training-entry-dialog/training-entry-dialog.component.ts
@@ -1,0 +1,1116 @@
+import { AsyncPipe } from '@angular/common';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  inject,
+  LOCALE_ID,
+  signal,
+} from '@angular/core';
+import { FormControl, ReactiveFormsModule } from '@angular/forms';
+import { toSignal } from '@angular/core/rxjs-interop';
+import { map, startWith } from 'rxjs';
+import { MatAutocompleteModule } from '@angular/material/autocomplete';
+import { MatButtonModule } from '@angular/material/button';
+import {
+  MAT_DIALOG_DATA,
+  MatDialogModule,
+  MatDialogRef,
+} from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatSelectModule } from '@angular/material/select';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { RouterLink } from '@angular/router';
+import {
+  appendLocalOffset,
+  COMPANION_BOUNDS,
+  EXERCISE_CATEGORIES,
+  exercisesByCategory,
+  ExerciseCategoryId,
+  ExerciseDefinition,
+  findExerciseDefinition,
+  findPushupTypeByLocalizedName,
+  findPushupTypeByStoredValue,
+  formatExerciseValue,
+  localizePushupType,
+  MeasurementType,
+  PUSHUP_TYPES,
+  PushupTypeInfo,
+} from '@pu-stats/models';
+
+/**
+ * Synthetic id for the implicit "Liegestütze (Standard)" exercise inside
+ * the pushup category. Pushups live in their own Firestore collection
+ * (`pushups`) and don't have a catalog `ExerciseDefinition`, so the
+ * exercise picker uses this stable sentinel to identify the row.
+ */
+const PUSHUP_EXERCISE_ID = 'pushup';
+
+const SECONDS_MAX = 59;
+
+/**
+ * Initial / edit-mode payload for {@link TrainingEntryDialogComponent}.
+ *
+ * `kind` discriminates which collection the entry lives in. In edit
+ * mode the picker rows for category and exercise are disabled — users
+ * cannot move an entry between collections after the fact.
+ */
+export interface TrainingEntryDialogData {
+  /** Discriminator. `'pushup'` populates pushup fields (variant + source);
+   *  `'exercise'` populates `exerciseId` + measurement-specific fields. */
+  kind: 'pushup' | 'exercise';
+  timestamp: string;
+  /** For `'pushup'`: total reps. For `'exercise'` reps measurement: total reps. */
+  reps?: number;
+  sets?: number[];
+  /** For `'exercise'` time / distance-time measurements. */
+  durationSec?: number;
+  /** For `'exercise'` distance-time measurement (meters). */
+  distanceM?: number;
+  /** For `'pushup'`: source (web / whatsapp / …). */
+  source?: string;
+  /** For `'pushup'`: variant id (e.g. `'standard'`, `'diamond'`). */
+  type?: string;
+  /** For `'exercise'`: catalog id (e.g. `'plank.standard'`). */
+  exerciseId?: string;
+  /** For `'exercise'`: variant id within the catalog definition. */
+  variantId?: string;
+}
+
+export interface TrainingEntryDialogResult {
+  kind: 'pushup' | 'exercise';
+  timestamp: string;
+  /** Total reps for reps-measurements; `0` otherwise. */
+  reps: number;
+  /** Per-set breakdown for reps-measurements; `[]` otherwise. */
+  sets: number[];
+  /** Set for `'time'` and `'distance-time'`. */
+  durationSec?: number;
+  /** Set for `'distance-time'`. */
+  distanceM?: number;
+  /** Pushup source (always set for `kind === 'pushup'`). */
+  source?: string;
+  /** Pushup variant id (always set for `kind === 'pushup'`). */
+  type?: string;
+  /** Catalog exercise id (always set for `kind === 'exercise'`). */
+  exerciseId?: string;
+  /** Catalog measurement (only set for `kind === 'exercise'`). */
+  measurement?: MeasurementType;
+  /**
+   * Variant id tri-state for `kind === 'exercise'`:
+   *   - non-empty `string`: set or keep this variant.
+   *   - `null`: clear an existing variant (caller maps to `deleteField()`).
+   *   - `undefined`: no change / no variant in create mode.
+   */
+  variantId?: string | null;
+}
+
+interface CategoryOption {
+  value: ExerciseCategoryId;
+  label: string;
+}
+
+interface ExerciseOption {
+  /** Catalog id, or {@link PUSHUP_EXERCISE_ID} for the pushup row. */
+  value: string;
+  label: string;
+  definition?: ExerciseDefinition;
+}
+
+interface PushupTypeOption {
+  value: string;
+  label: string;
+}
+
+/**
+ * Single dialog for entering / editing every training type the app
+ * supports. Two execution modes:
+ *
+ *   - **Pushup mode** (category = `'pushup'`): variant autocomplete +
+ *     source field + reps/sets list. Submits with `kind: 'pushup'` for
+ *     the legacy `pushups` Firestore collection.
+ *
+ *   - **Exercise mode** (every other category): exercise picker driven
+ *     by the catalog, then measurement-aware fields (reps/sets,
+ *     mm:ss, km + mm:ss). Submits with `kind: 'exercise'` for the
+ *     `exerciseEntries` collection.
+ *
+ * Edit mode (caller passes `data`) locks the category and exercise
+ * pickers — moving an entry between collections is not supported.
+ */
+@Component({
+  selector: 'app-training-entry-dialog',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [
+    AsyncPipe,
+    ReactiveFormsModule,
+    MatDialogModule,
+    MatButtonModule,
+    MatFormFieldModule,
+    MatIconModule,
+    MatInputModule,
+    MatSelectModule,
+    MatAutocompleteModule,
+    MatTooltipModule,
+    RouterLink,
+  ],
+  styles: [
+    `
+      mat-dialog-content {
+        display: grid;
+        gap: 10px;
+      }
+      mat-form-field:first-of-type {
+        margin-top: 8px;
+      }
+      .set-row {
+        display: flex;
+        align-items: center;
+        gap: 8px;
+        animation: clone-set 300ms cubic-bezier(0.4, 0, 0.2, 1) both;
+        transform-origin: top center;
+      }
+      .set-row mat-form-field {
+        flex: 1;
+      }
+      .duration-row {
+        display: flex;
+        gap: 8px;
+      }
+      .duration-row mat-form-field {
+        flex: 1;
+      }
+      @keyframes clone-set {
+        from {
+          opacity: 0;
+          max-height: 0;
+          transform: scaleY(0.3) translateY(-8px);
+        }
+        to {
+          opacity: 1;
+          max-height: 80px;
+          transform: scaleY(1) translateY(0);
+        }
+      }
+      .total-reps {
+        font-size: 13px;
+        color: var(--mat-sys-on-surface-variant);
+        text-align: end;
+      }
+      .type-row {
+        display: flex;
+        align-items: flex-start;
+        gap: 4px;
+      }
+      .type-row mat-form-field {
+        flex: 1;
+      }
+      .type-help {
+        margin-top: 12px;
+      }
+      .type-option {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        gap: 12px;
+        width: 100%;
+      }
+      .type-option-info {
+        font-size: 18px;
+        width: 18px;
+        height: 18px;
+        line-height: 18px;
+        color: var(--mat-sys-on-surface-variant);
+        cursor: help;
+        flex: 0 0 auto;
+      }
+    `,
+  ],
+  template: `
+    <h2 mat-dialog-title>
+      @if (isEditMode) {
+        <span i18n="@@editDialogTitle">Eintrag bearbeiten</span>
+      } @else {
+        <span i18n="@@trainingEntryDialog.title">Trainingseintrag anlegen</span>
+      }
+    </h2>
+
+    <mat-dialog-content>
+      <mat-form-field appearance="outline">
+        <mat-label i18n="@@trainingEntryDialog.category">Kategorie</mat-label>
+        <mat-select
+          [value]="category()"
+          (valueChange)="onCategoryChange($event)"
+          [disabled]="isEditMode"
+          data-testid="training-entry-category"
+        >
+          @for (option of categoryOptions; track option.value) {
+            <mat-option [value]="option.value">{{ option.label }}</mat-option>
+          }
+        </mat-select>
+      </mat-form-field>
+
+      @if (showExercisePicker()) {
+        <mat-form-field appearance="outline">
+          <mat-label i18n="@@trainingEntryDialog.exercise">Übung</mat-label>
+          <mat-select
+            [value]="exerciseId()"
+            (valueChange)="exerciseId.set($event)"
+            [disabled]="isEditMode"
+            data-testid="training-entry-exercise"
+          >
+            @for (option of exerciseOptions(); track option.value) {
+              <mat-option [value]="option.value">{{ option.label }}</mat-option>
+            }
+          </mat-select>
+        </mat-form-field>
+      }
+
+      <mat-form-field appearance="outline">
+        <mat-label i18n="@@timestampLabel">Zeitpunkt</mat-label>
+        <input
+          matInput
+          type="datetime-local"
+          [value]="timestamp()"
+          (input)="timestamp.set(asValue($event))"
+          required
+        />
+      </mat-form-field>
+
+      @if (mode() === 'pushup') {
+        <div class="type-row">
+          <mat-form-field appearance="outline">
+            <mat-label i18n="@@typeLabel">Typ</mat-label>
+            <input
+              type="text"
+              matInput
+              [formControl]="pushupTypeControl"
+              [matAutocomplete]="typeAuto"
+              placeholder="Auswählen oder eintippen"
+              i18n-placeholder="@@typePlaceholder"
+            />
+            <mat-autocomplete
+              #typeAuto="matAutocomplete"
+              [displayWith]="displayPushupType"
+            >
+              @for (
+                option of filteredPushupTypeOptions$ | async;
+                track option.value
+              ) {
+                <mat-option [value]="option.value">
+                  <span class="type-option">
+                    <span>{{ option.label }}</span>
+                    @if (pushupTooltipFor(option.value); as info) {
+                      <mat-icon
+                        class="type-option-info"
+                        [matTooltip]="info"
+                        matTooltipPosition="right"
+                        [attr.aria-label]="info"
+                        role="img"
+                        (click)="$event.stopPropagation()"
+                        >info_outline</mat-icon
+                      >
+                    }
+                  </span>
+                </mat-option>
+              }
+            </mat-autocomplete>
+          </mat-form-field>
+          <a
+            mat-icon-button
+            class="type-help"
+            [routerLink]="['/wiki/liegestuetz-typen']"
+            [queryParams]="pushupWikiQueryParams()"
+            [matTooltip]="pushupWikiTooltip()"
+            matTooltipPosition="left"
+            mat-dialog-close
+            i18n-aria-label="@@typeWikiLinkAria"
+            aria-label="Anleitung zum Liegestütztyp öffnen"
+          >
+            <mat-icon>help_outline</mat-icon>
+          </a>
+        </div>
+      } @else if (showVariantPicker()) {
+        <mat-form-field appearance="outline">
+          <mat-label i18n="@@entryDialog.variant">Variante</mat-label>
+          <mat-select [formControl]="variantControl">
+            <mat-option [value]="''" i18n="@@entryDialog.variantNone"
+              >—</mat-option
+            >
+            @for (variant of currentDefinition()?.variants; track variant.id) {
+              <mat-option [value]="variant.id">{{
+                exerciseVariantLabel(variant.id)
+              }}</mat-option>
+            }
+          </mat-select>
+        </mat-form-field>
+      }
+
+      @if (isDistanceTimeMeasurement()) {
+        <mat-form-field appearance="outline">
+          <mat-label i18n="@@entryDialog.distance">Distanz (km)</mat-label>
+          <input
+            matInput
+            type="number"
+            inputmode="decimal"
+            placeholder="5.00"
+            [min]="minKm()"
+            [max]="maxKm()"
+            step="0.01"
+            [value]="distanceInput()"
+            (input)="distanceInput.set(asValue($event))"
+            required
+          />
+        </mat-form-field>
+        <div class="duration-row">
+          <mat-form-field appearance="outline">
+            <mat-label i18n="@@entryDialog.durationMinutes">Minuten</mat-label>
+            <input
+              matInput
+              type="number"
+              inputmode="numeric"
+              placeholder="25"
+              min="0"
+              step="1"
+              [value]="durationMinutesInput()"
+              (input)="durationMinutesInput.set(asValue($event))"
+            />
+          </mat-form-field>
+          <mat-form-field appearance="outline">
+            <mat-label i18n="@@entryDialog.durationSeconds">Sekunden</mat-label>
+            <input
+              matInput
+              type="number"
+              inputmode="numeric"
+              placeholder="0"
+              min="0"
+              [max]="secondsMax"
+              step="1"
+              [value]="durationSecondsInput()"
+              (input)="durationSecondsInput.set(asValue($event))"
+            />
+          </mat-form-field>
+        </div>
+      } @else if (isTimeMeasurement()) {
+        <div class="duration-row">
+          <mat-form-field appearance="outline">
+            <mat-label i18n="@@entryDialog.durationMinutes">Minuten</mat-label>
+            <input
+              matInput
+              type="number"
+              inputmode="numeric"
+              placeholder="1"
+              min="0"
+              step="1"
+              [value]="durationMinutesInput()"
+              (input)="durationMinutesInput.set(asValue($event))"
+            />
+          </mat-form-field>
+          <mat-form-field appearance="outline">
+            <mat-label i18n="@@entryDialog.durationSeconds">Sekunden</mat-label>
+            <input
+              matInput
+              type="number"
+              inputmode="numeric"
+              placeholder="30"
+              min="0"
+              [max]="secondsMax"
+              step="1"
+              [value]="durationSecondsInput()"
+              (input)="durationSecondsInput.set(asValue($event))"
+            />
+          </mat-form-field>
+        </div>
+      } @else {
+        @for (set of sets(); track $index) {
+          <div class="set-row">
+            <mat-form-field appearance="outline">
+              @if (hasMultipleSets()) {
+                <mat-label i18n="@@setLabel">Set {{ $index + 1 }}</mat-label>
+              } @else {
+                <mat-label i18n="@@repsLabel">Reps</mat-label>
+              }
+              <input
+                matInput
+                type="number"
+                [min]="0"
+                [max]="repsMax()"
+                step="1"
+                [value]="set"
+                (input)="updateSet($index, asValue($event))"
+                required
+              />
+            </mat-form-field>
+            @if (hasMultipleSets()) {
+              <button
+                type="button"
+                mat-icon-button
+                (click)="removeSet($index)"
+                i18n-aria-label="@@removeSetAria"
+                aria-label="Set entfernen"
+              >
+                <mat-icon>remove_circle_outline</mat-icon>
+              </button>
+            }
+            @if ($last) {
+              <button
+                type="button"
+                mat-icon-button
+                (click)="addSet()"
+                i18n-aria-label="@@addSetAria"
+                aria-label="Set hinzufügen"
+                i18n-matTooltip="@@addSetTooltip"
+                matTooltip="Weiteres Set hinzufügen"
+              >
+                <mat-icon>add_circle_outline</mat-icon>
+              </button>
+            }
+          </div>
+        }
+        @if (hasMultipleSets()) {
+          <div class="total-reps" i18n="@@totalReps">
+            Gesamt: {{ totalReps() }} Reps
+          </div>
+        }
+      }
+
+      @if (overCap()) {
+        <div class="total-reps">
+          @if (overCapKind() === 'duration') {
+            <span i18n="@@entryDialog.overCapDuration"
+              >Maximum-Dauer: {{ formattedDurationMax() }}</span
+            >
+          } @else if (overCapKind() === 'distance' || isTimeMeasurement()) {
+            <span i18n="@@entryDialog.overCapTime"
+              >Maximum: {{ formattedMax() }}</span
+            >
+          } @else {
+            <span i18n="@@entryDialog.overCap"
+              >Maximum: {{ repsMax() }} pro Eintrag</span
+            >
+          }
+        </div>
+      }
+
+      @if (mode() === 'pushup') {
+        <mat-form-field appearance="outline">
+          <mat-label i18n="@@sourceLabel">Quelle</mat-label>
+          <input
+            type="text"
+            matInput
+            [formControl]="sourceControl"
+            [matAutocomplete]="sourceAuto"
+            placeholder="web / whatsapp / eigene Quelle"
+            i18n-placeholder="@@sourcePlaceholder"
+          />
+          <mat-autocomplete #sourceAuto="matAutocomplete">
+            @for (option of filteredSourceOptions$ | async; track option) {
+              <mat-option [value]="option">{{ option }}</mat-option>
+            }
+          </mat-autocomplete>
+        </mat-form-field>
+      }
+    </mat-dialog-content>
+
+    <mat-dialog-actions align="end">
+      <button type="button" mat-button mat-dialog-close i18n="@@cancel">
+        Abbrechen
+      </button>
+      <button
+        type="button"
+        mat-flat-button
+        [disabled]="!canSubmit()"
+        (click)="submit()"
+        i18n="@@saveEntry"
+        data-testid="training-entry-submit"
+      >
+        Speichern
+      </button>
+    </mat-dialog-actions>
+  `,
+})
+export class TrainingEntryDialogComponent {
+  private readonly dialogRef = inject(
+    MatDialogRef<TrainingEntryDialogComponent>
+  );
+  private readonly data = inject<TrainingEntryDialogData | null>(
+    MAT_DIALOG_DATA,
+    {
+      optional: true,
+    }
+  );
+  private readonly locale = inject(LOCALE_ID) as string;
+
+  readonly isEditMode = !!this.data;
+
+  /**
+   * Cap pushup reps at the same per-entry ceiling other catalog
+   * exercises use. Pushups currently have no per-record schema cap,
+   * but {@link COMPANION_BOUNDS}.reps gives us a reasonable shared
+   * upper bound (500) that prevents typo-driven absurd entries.
+   */
+  private readonly pushupRepsMax = COMPANION_BOUNDS.reps.max;
+
+  readonly categoryOptions: ReadonlyArray<CategoryOption> =
+    buildCategoryOptions();
+
+  // ----- selection state ---------------------------------------------------
+
+  readonly category = signal<ExerciseCategoryId>(this.initialCategory());
+
+  /**
+   * Catalog exercise id when category !== 'pushup'. Defaults to the
+   * first exercise of the chosen category. For the pushup category the
+   * value is the synthetic {@link PUSHUP_EXERCISE_ID} so the picker can
+   * show a single entry without a real catalog row.
+   */
+  readonly exerciseId = signal<string>(this.initialExerciseId());
+
+  readonly mode = computed<'pushup' | 'exercise'>(() =>
+    this.category() === 'pushup' ? 'pushup' : 'exercise'
+  );
+
+  readonly exerciseOptions = computed<ExerciseOption[]>(() => {
+    const cat = this.category();
+    if (cat === 'pushup') {
+      return [
+        {
+          value: PUSHUP_EXERCISE_ID,
+          label: $localize`:@@trainingEntryDialog.pushup.exercise:Liegestütze`,
+        },
+      ];
+    }
+    const defs = exercisesByCategory().get(cat) ?? [];
+    return defs.map((def) => ({
+      value: def.id,
+      label: this.exerciseLabel(def.id),
+      definition: def,
+    }));
+  });
+
+  readonly showExercisePicker = computed(
+    () => this.mode() === 'exercise' && this.exerciseOptions().length > 0
+  );
+
+  readonly currentDefinition = computed<ExerciseDefinition | null>(() => {
+    if (this.mode() !== 'exercise') return null;
+    return findExerciseDefinition(this.exerciseId());
+  });
+
+  // ----- timestamp & set state --------------------------------------------
+
+  private readonly originalTimestamp = this.data?.timestamp ?? null;
+
+  readonly timestamp = signal(
+    this.data ? this.data.timestamp.slice(0, 16) : this.defaultDateTimeLocal()
+  );
+
+  readonly sets = signal<number[]>(this.initialSets());
+
+  readonly hasMultipleSets = computed(() => this.sets().length > 1);
+
+  readonly totalReps = computed(() =>
+    this.sets().reduce((sum, s) => sum + (s > 0 ? s : 0), 0)
+  );
+
+  /** Active reps cap — catalog `def.max` for exercise mode, fixed for pushup. */
+  readonly repsMax = computed(() => {
+    if (this.mode() === 'pushup') return this.pushupRepsMax;
+    return this.currentDefinition()?.max ?? this.pushupRepsMax;
+  });
+
+  // ----- duration / distance ----------------------------------------------
+
+  /**
+   * Split duration input — minutes + seconds — for time-measurement
+   * exercises (plank) and the duration companion of distance-time
+   * exercises (cardio.running). Stored as strings so an empty field is
+   * representable; combined into integer seconds via {@link durationSec}.
+   */
+  private readonly initialDurationParts = splitDurationParts(
+    this.data?.durationSec
+  );
+
+  readonly durationMinutesInput = signal(this.initialDurationParts.minutes);
+  readonly durationSecondsInput = signal(this.initialDurationParts.seconds);
+
+  readonly durationSec = computed(() =>
+    parseDurationFromParts(
+      this.durationMinutesInput(),
+      this.durationSecondsInput()
+    )
+  );
+
+  readonly secondsMax = SECONDS_MAX;
+
+  readonly distanceInput = signal(
+    this.data?.distanceM !== undefined ? formatKmInput(this.data.distanceM) : ''
+  );
+
+  readonly distanceM = computed(() => parseKmToMeters(this.distanceInput()));
+
+  readonly minKm = computed(() => (this.currentDefinition()?.min ?? 0) / 1000);
+  readonly maxKm = computed(() => (this.currentDefinition()?.max ?? 0) / 1000);
+
+  readonly isTimeMeasurement = computed(
+    () => this.currentDefinition()?.measurement === 'time'
+  );
+
+  readonly isDistanceTimeMeasurement = computed(
+    () => this.currentDefinition()?.measurement === 'distance-time'
+  );
+
+  // ----- exercise variant picker ------------------------------------------
+
+  readonly variantControl = new FormControl<string>(
+    this.data?.kind === 'exercise' ? (this.data.variantId ?? '') : '',
+    { nonNullable: true }
+  );
+
+  readonly showVariantPicker = computed(() => {
+    if (this.mode() !== 'exercise') return false;
+    return (this.currentDefinition()?.variants?.length ?? 0) > 0;
+  });
+
+  // ----- pushup type & source ---------------------------------------------
+
+  readonly pushupTypeControl = new FormControl<string>(
+    this.data?.kind === 'pushup' ? this.data.type || 'standard' : 'standard',
+    { nonNullable: true }
+  );
+
+  readonly sourceControl = new FormControl<string>(
+    this.normalizeSource(
+      this.data?.kind === 'pushup' ? this.data.source || 'web' : 'web'
+    ),
+    { nonNullable: true }
+  );
+
+  private readonly pushupTypeOptions: ReadonlyArray<PushupTypeOption> =
+    PUSHUP_TYPES.map((t) => ({
+      value: t.id,
+      label: localizePushupType(t, this.locale).name,
+    }));
+
+  private readonly sourceOptions = ['web', 'whatsapp'];
+
+  readonly filteredPushupTypeOptions$ =
+    this.pushupTypeControl.valueChanges.pipe(
+      startWith(this.pushupTypeControl.value),
+      map((value) => this.filterPushupTypeOptions(value))
+    );
+
+  readonly filteredSourceOptions$ = this.sourceControl.valueChanges.pipe(
+    startWith(this.sourceControl.value),
+    map((value) => this.filterOptions(value, this.sourceOptions))
+  );
+
+  readonly displayPushupType = (value: string | null | undefined): string => {
+    if (!value) return '';
+    const match = findPushupTypeByStoredValue(value);
+    return match ? localizePushupType(match, this.locale).name : value;
+  };
+
+  private readonly pushupTypeValue = toSignal(
+    this.pushupTypeControl.valueChanges.pipe(
+      startWith(this.pushupTypeControl.value)
+    ),
+    { initialValue: this.pushupTypeControl.value }
+  );
+
+  readonly pushupWikiQueryParams = computed(() => {
+    const match = this.resolvePushupType(this.pushupTypeValue());
+    return match ? { type: match.slug } : {};
+  });
+
+  readonly pushupWikiTooltip = computed(() => {
+    const match = this.resolvePushupType(this.pushupTypeValue());
+    if (!match) {
+      return $localize`:@@typeWikiLinkTooltip.generic:Anleitung zu Liegestütztypen öffnen`;
+    }
+    const template = $localize`:@@typeWikiLinkTooltip.specific:Anleitung öffnen`;
+    return `${template}: ${localizePushupType(match, this.locale).name}`;
+  });
+
+  pushupTooltipFor(label: string): string {
+    const type = this.resolvePushupType(label);
+    return type ? localizePushupType(type, this.locale).summary : '';
+  }
+
+  // ----- caps & submit gating ---------------------------------------------
+
+  readonly formattedMax = computed(() => {
+    const def = this.currentDefinition();
+    if (!def) return String(this.repsMax());
+    return formatExerciseValue(def.max, def.unit);
+  });
+
+  readonly formattedDurationMax = computed(() =>
+    formatExerciseValue(COMPANION_BOUNDS.durationSec.max, 's')
+  );
+
+  readonly overCapKind = computed<'distance' | 'duration' | 'value' | null>(
+    () => {
+      if (this.isDistanceTimeMeasurement()) {
+        const def = this.currentDefinition();
+        if (!def) return null;
+        const m = this.distanceM();
+        const sec = this.durationSec();
+        if (m !== null && m > def.max) return 'distance';
+        if (sec !== null && sec > COMPANION_BOUNDS.durationSec.max)
+          return 'duration';
+        return null;
+      }
+      if (this.isTimeMeasurement()) {
+        const def = this.currentDefinition();
+        if (!def) return null;
+        const sec = this.durationSec();
+        return sec !== null && sec > def.max ? 'value' : null;
+      }
+      return this.totalReps() > this.repsMax() ? 'value' : null;
+    }
+  );
+
+  readonly overCap = computed(() => this.overCapKind() !== null);
+
+  readonly canSubmit = computed(() => {
+    if (this.timestamp().length === 0) return false;
+    if (this.mode() === 'pushup') {
+      return this.totalReps() > 0 && !this.overCap();
+    }
+    const def = this.currentDefinition();
+    if (!def) return false;
+    if (this.isDistanceTimeMeasurement()) {
+      const m = this.distanceM();
+      const sec = this.durationSec();
+      return (
+        m !== null && m >= def.min && sec !== null && sec > 0 && !this.overCap()
+      );
+    }
+    if (this.isTimeMeasurement()) {
+      const sec = this.durationSec();
+      return sec !== null && sec > 0 && sec >= def.min && !this.overCap();
+    }
+    return this.totalReps() >= def.min && !this.overCap();
+  });
+
+  // ----- handlers ---------------------------------------------------------
+
+  onCategoryChange(next: ExerciseCategoryId): void {
+    if (this.isEditMode) return;
+    this.category.set(next);
+    // Reset exercise to the first option of the new category. Reset
+    // measurement-specific inputs so a stale value from a previously
+    // selected exercise can't leak through to the new one.
+    if (next === 'pushup') {
+      this.exerciseId.set(PUSHUP_EXERCISE_ID);
+    } else {
+      const defs = exercisesByCategory().get(next) ?? [];
+      this.exerciseId.set(defs[0]?.id ?? '');
+    }
+    this.sets.set([0]);
+    this.durationMinutesInput.set('');
+    this.durationSecondsInput.set('');
+    this.distanceInput.set('');
+    this.variantControl.setValue('');
+  }
+
+  exerciseVariantLabel(variantId: string): string {
+    return variantId;
+  }
+
+  exerciseLabel(id: string): string {
+    return EXERCISE_NAMES[id] ?? id;
+  }
+
+  addSet(): void {
+    const last = this.sets()[this.sets().length - 1] ?? 0;
+    this.sets.update((s) => [...s, last > 0 ? last : 0]);
+  }
+
+  removeSet(index: number): void {
+    this.sets.update((s) => {
+      const next = s.filter((_, i) => i !== index);
+      return next.length === 0 ? [0] : next;
+    });
+  }
+
+  updateSet(index: number, value: string): void {
+    const parsed = Number(value);
+    const finite = Number.isFinite(parsed)
+      ? Math.max(0, Math.floor(parsed))
+      : 0;
+    const clamped = Math.min(finite, this.repsMax());
+    this.sets.update((s) => s.map((v, i) => (i === index ? clamped : v)));
+  }
+
+  asValue(event: Event): string {
+    return (event.target as HTMLInputElement).value;
+  }
+
+  submit(): void {
+    if (!this.canSubmit()) return;
+
+    const defaultLocal = this.originalTimestamp?.slice(0, 16) ?? '';
+    const timestamp =
+      this.originalTimestamp && this.timestamp() === defaultLocal
+        ? this.originalTimestamp
+        : appendLocalOffset(this.timestamp());
+
+    if (this.mode() === 'pushup') {
+      const validSets = this.sets().filter((s) => s > 0);
+      const reps = validSets.reduce((sum, s) => sum + s, 0);
+      if (reps <= 0) return;
+
+      const rawType = (this.pushupTypeControl.value || '').trim() || 'standard';
+      const matched = this.resolvePushupType(rawType);
+      const type = matched ? matched.id : rawType;
+      const source = this.normalizeSource(
+        (this.sourceControl.value || '').trim() || 'web'
+      );
+
+      const result: TrainingEntryDialogResult = {
+        kind: 'pushup',
+        timestamp,
+        reps,
+        sets: validSets,
+        source,
+        type,
+      };
+      this.dialogRef.close(result);
+      return;
+    }
+
+    const def = this.currentDefinition();
+    if (!def) return;
+
+    // Variant tri-state preservation — only emit when changed so we
+    // don't race against concurrent edits with stale data.
+    const variantId = this.variantControl.value.trim();
+    const initialVariantId =
+      this.data?.kind === 'exercise' ? (this.data.variantId ?? '') : '';
+    const variantPatch: { variantId?: string | null } =
+      variantId === initialVariantId
+        ? {}
+        : variantId
+          ? { variantId }
+          : initialVariantId
+            ? { variantId: null }
+            : {};
+
+    const measurement = def.measurement;
+
+    if (measurement === 'time') {
+      const sec = this.durationSec();
+      if (sec === null || sec <= 0) return;
+      const result: TrainingEntryDialogResult = {
+        kind: 'exercise',
+        exerciseId: def.id,
+        measurement,
+        ...variantPatch,
+        timestamp,
+        reps: 0,
+        sets: [],
+        durationSec: sec,
+      };
+      this.dialogRef.close(result);
+      return;
+    }
+
+    if (measurement === 'distance-time') {
+      const m = this.distanceM();
+      const sec = this.durationSec();
+      if (m === null || m <= 0 || sec === null || sec <= 0) return;
+      const result: TrainingEntryDialogResult = {
+        kind: 'exercise',
+        exerciseId: def.id,
+        measurement,
+        ...variantPatch,
+        timestamp,
+        reps: 0,
+        sets: [],
+        distanceM: m,
+        durationSec: sec,
+      };
+      this.dialogRef.close(result);
+      return;
+    }
+
+    const validSets = this.sets().filter((s) => s > 0);
+    const reps = validSets.reduce((sum, s) => sum + s, 0);
+    if (reps <= 0) return;
+
+    const result: TrainingEntryDialogResult = {
+      kind: 'exercise',
+      exerciseId: def.id,
+      measurement,
+      ...variantPatch,
+      timestamp,
+      reps,
+      sets: validSets,
+    };
+    this.dialogRef.close(result);
+  }
+
+  // ----- helpers ----------------------------------------------------------
+
+  private initialCategory(): ExerciseCategoryId {
+    if (!this.data) return 'pushup';
+    if (this.data.kind === 'pushup') return 'pushup';
+    const def = findExerciseDefinition(this.data.exerciseId);
+    return def?.categoryId ?? 'pushup';
+  }
+
+  private initialExerciseId(): string {
+    if (!this.data) return PUSHUP_EXERCISE_ID;
+    if (this.data.kind === 'pushup') return PUSHUP_EXERCISE_ID;
+    return this.data.exerciseId ?? '';
+  }
+
+  private initialSets(): number[] {
+    if (!this.data) return [0];
+    if (this.data.sets?.length) return [...this.data.sets];
+    if (this.data.reps !== undefined && this.data.reps > 0)
+      return [this.data.reps];
+    return [0];
+  }
+
+  private resolvePushupType(
+    value: string | null | undefined
+  ): PushupTypeInfo | null {
+    return findPushupTypeByLocalizedName(value, this.locale);
+  }
+
+  private normalizeSource(value: string): string {
+    const v = (value || '').trim();
+    if (!v) return 'web';
+    if (v === 'wa') return 'whatsapp';
+    return v;
+  }
+
+  private filterOptions(
+    value: string | null | undefined,
+    options: ReadonlyArray<string>
+  ): string[] {
+    const needle = (value ?? '').toLowerCase().trim();
+    if (!needle) return [...options];
+    if (options.some((opt) => opt.toLowerCase() === needle))
+      return [...options];
+    return options.filter((opt) => opt.toLowerCase().includes(needle));
+  }
+
+  private filterPushupTypeOptions(
+    value: string | null | undefined
+  ): PushupTypeOption[] {
+    const needle = (value ?? '').toLowerCase().trim();
+    if (!needle) return [...this.pushupTypeOptions];
+    const exact = this.pushupTypeOptions.some(
+      (opt) =>
+        opt.value.toLowerCase() === needle || opt.label.toLowerCase() === needle
+    );
+    if (exact) return [...this.pushupTypeOptions];
+    return this.pushupTypeOptions.filter(
+      (opt) =>
+        opt.label.toLowerCase().includes(needle) ||
+        opt.value.toLowerCase().includes(needle)
+    );
+  }
+
+  private defaultDateTimeLocal(): string {
+    const now = new Date();
+    const pad = (n: number) => String(n).padStart(2, '0');
+    return `${now.getFullYear()}-${pad(now.getMonth() + 1)}-${pad(
+      now.getDate()
+    )}T${pad(now.getHours())}:${pad(now.getMinutes())}`;
+  }
+}
+
+const CATEGORY_NAMES: Record<ExerciseCategoryId, () => string> = {
+  pushup: () => $localize`:@@exercise.category.pushup:Liegestütze`,
+  abs: () => $localize`:@@exercise.category.abs:Bauch`,
+  legs: () => $localize`:@@exercise.category.legs:Beine`,
+  plank: () => $localize`:@@exercise.category.plank:Plank`,
+  cardio: () => $localize`:@@exercise.category.cardio:Ausdauer`,
+  // strength + mobility are catalog-known but have no exercises yet —
+  // leave them out of the dialog until they ship a definition.
+  strength: () => 'Kraft',
+  mobility: () => 'Mobility',
+};
+
+const EXERCISE_NAMES: Record<string, string> = {
+  'abs.situps': $localize`:@@exercise.abs.situps.name:Sit-ups`,
+  'legs.squats': $localize`:@@exercise.legs.squats.name:Kniebeugen`,
+  'plank.standard': $localize`:@@exercise.plank.standard.name:Plank`,
+  'cardio.running': $localize`:@@exercise.cardio.running.name:Laufen`,
+};
+
+function buildCategoryOptions(): ReadonlyArray<CategoryOption> {
+  // Filter to categories that actually have something to log in this
+  // dialog: pushup (always — handled separately) plus any catalog
+  // category with at least one ExerciseDefinition.
+  const byCategory = exercisesByCategory();
+  return EXERCISE_CATEGORIES.filter(
+    (cat) => cat.id === 'pushup' || (byCategory.get(cat.id)?.length ?? 0) > 0
+  ).map((cat) => ({
+    value: cat.id,
+    label: (CATEGORY_NAMES[cat.id] ?? (() => cat.id))(),
+  }));
+}
+
+/**
+ * Combine the two number-input parts into total integer seconds.
+ * Returns `null` when both parts are empty or any field is invalid, so
+ * the caller can disable submit instead of writing `NaN`/`0` to
+ * Firestore. Either part empty is treated as 0 — typing only minutes or
+ * only seconds is a valid shortcut.
+ */
+function parseDurationFromParts(
+  minutesInput: string,
+  secondsInput: string
+): number | null {
+  const minTrim = minutesInput.trim();
+  const secTrim = secondsInput.trim();
+  if (minTrim === '' && secTrim === '') return null;
+
+  const minutes = minTrim === '' ? 0 : Number(minTrim);
+  const seconds = secTrim === '' ? 0 : Number(secTrim);
+  // Reject fractional input (e.g. "1.9") so a paste-into-the-number-field
+  // doesn't silently truncate to a shorter saved duration.
+  if (!Number.isInteger(minutes) || !Number.isInteger(seconds)) return null;
+  if (minutes < 0 || seconds < 0) return null;
+  if (seconds > SECONDS_MAX) return null;
+
+  return minutes * 60 + seconds;
+}
+
+/**
+ * Split a stored total-seconds value into the two strings shown in the
+ * minutes / seconds inputs. Empty strings when no initial value, so the
+ * fields render blank in create mode.
+ */
+function splitDurationParts(totalSec: number | undefined): {
+  minutes: string;
+  seconds: string;
+} {
+  if (totalSec === undefined || !Number.isFinite(totalSec) || totalSec <= 0) {
+    return { minutes: '', seconds: '' };
+  }
+  const m = Math.floor(totalSec / 60);
+  const s = Math.floor(totalSec % 60);
+  return { minutes: String(m), seconds: String(s) };
+}
+
+function parseKmToMeters(input: string): number | null {
+  const trimmed = input.trim().replace(',', '.');
+  if (!trimmed) return null;
+  const km = Number(trimmed);
+  if (!Number.isFinite(km) || km <= 0) return null;
+  return Math.round(km * 1000);
+}
+
+function formatKmInput(distanceM: number): string {
+  if (!Number.isFinite(distanceM) || distanceM <= 0) return '';
+  return (distanceM / 1000).toFixed(2);
+}
