+++ web/src/locale/messages.el.xlf
@@ -7426,5 +7426,29 @@
         <target>Προσαρμογή χειροκίνητων στόχων</target>
       </segment>
     </unit>
+    <unit id="trainingEntryDialog.title">
+      <segment state="translated">
+        <source>Trainingseintrag anlegen</source>
+        <target>Προσθήκη καταχώρησης προπόνησης</target>
+      </segment>
+    </unit>
+    <unit id="trainingEntryDialog.category">
+      <segment state="translated">
+        <source>Kategorie</source>
+        <target>Κατηγορία</target>
+      </segment>
+    </unit>
+    <unit id="trainingEntryDialog.exercise">
+      <segment state="translated">
+        <source>Übung</source>
+        <target>Άσκηση</target>
+      </segment>
+    </unit>
+    <unit id="trainingEntryDialog.pushup.exercise">
+      <segment state="translated">
+        <source>Liegestütze</source>
+        <target>Κάμψεις</target>
+      </segment>
+    </unit>
   </file>
 </xliff>
