+++ web/src/locale/messages.nl.xlf
@@ -7426,5 +7426,29 @@
         <target>Handmatige doelen aanpassen</target>
       </segment>
     </unit>
+    <unit id="trainingEntryDialog.title">
+      <segment state="translated">
+        <source>Trainingseintrag anlegen</source>
+        <target>Trainingsvermelding toevoegen</target>
+      </segment>
+    </unit>
+    <unit id="trainingEntryDialog.category">
+      <segment state="translated">
+        <source>Kategorie</source>
+        <target>Categorie</target>
+      </segment>
+    </unit>
+    <unit id="trainingEntryDialog.exercise">
+      <segment state="translated">
+        <source>Übung</source>
+        <target>Oefening</target>
+      </segment>
+    </unit>
+    <unit id="trainingEntryDialog.pushup.exercise">
+      <segment state="translated">
+        <source>Liegestütze</source>
+        <target>Push-ups</target>
+      </segment>
+    </unit>
   </file>
 </xliff>
